requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
pycparser==2.22
PyMySQL==1.1.1
PyPDF2==3.0.1
python-docx==1.1.2
python-magic==0.4.27
SQLAlchemy==2.0.35
typing_extensions==4.12.2
Werkzeug==3.0.4
